services/python-rag/requirements.txt
# PadhAI-Dost Python RAG Service

fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
python-multipart>=0.0.9

# Retrieval
langchain>=0.1.0
langchain-community>=0.0.20

# Embedding
transformers>=4.36.0
torch>=2.1.0
sentence-transformers>=2.3.0

# Vector Store
chromadb>=0.4.22

# Document Processing
PyPDF2>=3.0.0
Pillow>=10.0.0
pytesseract>=0.3.10
pdfplumber>=0.10.0

# Optional: PDF to image for Tesseract OCR on scanned PDFs
pdf2image>=1.16.0
